Make `fixed_t` `long` instead of `int`.
You'd think one of the most ported pieces of software in history
wouldn't assume `int` to be 32-bit but it does. Unbelieveable.

+++ linuxdoom-1.10/m_fixed.h
@@ -29,9 +29,9 @@
 // Fixed point, 32bit as 16.16.
 //
 #define FRACBITS		16
-#define FRACUNIT		(1<<FRACBITS)
+#define FRACUNIT		(1l<<FRACBITS)
 
-typedef int fixed_t;
+typedef long fixed_t;
 
 fixed_t FixedMul	(fixed_t a, fixed_t b);
 fixed_t FixedDiv	(fixed_t a, fixed_t b);
